Finally build working, with dirty can hack will wait for developer's response meanwhile good enough for rehosting

+++ libraries/AP_Filesystem/AP_Filesystem_FATFS.cpp
@@ -525,7 +525,7 @@ int AP_Filesystem_FATFS::fsync(int fileno)
     return 0;
 }
 
-off_t AP_Filesystem_FATFS::lseek(int fileno, off_t position, int whence)
+int32_t AP_Filesystem_FATFS::lseek(int fileno, int32_t position, int whence)
 {
     FRESULT res;
     FIL *fh;
